+++ test/support/messenger_handler.ex
@@ -1,7 +1,7 @@
 defmodule Solvent.MessengerHandler do
   use Solvent.Subscriber,
     id: "messenger subscriber",
-    match_type: ~r/modulesubscribe\..*/
+    match_type: "modulesubscribe.published"
 
   require Logger